refactor:#34 유틸함수 리팩터링, 코드 분리
